refactor: move mocket.inject to mocket.core.inject
